Files Added
Files to repository added

The files for the project STM32F4 with USB
peripheral drivers and
testing sending function are for the first time
added.
In this files the STM32F4 is functioning with following
features:
- HSE Clock
- USB drivers for device configuration an
FS-mode
A CubeMX project is also added so that it is posible to take a
look of
the configurations.

The .gitignore file is modified in order some files not to be commited